+++ src/components/Carousel/Carousel.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import cx from 'classnames';
 
-import { Button, ButtonIcon } from '../../';
+import { IconButton } from '../Button';
 
 class Carousel extends Component {
   state = {
@@ -11,7 +11,6 @@ class Carousel extends Component {
   }
 
   autoPlayId = null // eslint-disable-line react/sort-comp
-  autoPlayFlavors = ['icon-border-filled', 'icon-x-small']
 
   componentWillMount() {
     const { autoPlayActive, children } = this.props;
@@ -147,15 +146,14 @@ class Carousel extends Component {
 
     return (
       <span className="slds-carousel__autoplay">
-        <Button
-          flavor={this.autoPlayFlavors}
+        <IconButton
+          icon={icon}
+          border="filled"
+          sprite="utility"
           onClick={this.toggleAutoPlay}
           size="x-small"
           title={assistiveText}
-        >
-          <ButtonIcon sprite="utility" icon={icon} />
-          <span className="slds-assistive-text">{assistiveText}</span>
-        </Button>
+        />
       </span>
     );
   }