+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 5.0.0
+
+This version is compatible with SLDS 2.5.2 aka `Spring '18` 🌸
+
+* _Tabs:_ **Breaking** Reimplement `Tabs`, `ControlledTabs`
+* _Tabs_: Support keyboard interactions for `Tabs`
+* _Button:_ **Breaking** Reimplement `Button`s, add `IconButton`, `StatefulButton`, `StatefulIconButton` components
+* _Modal:_ **Breaking** Reimplement `Modal`
+* _Modal:_ Trap focus inside `Modal`, support keyboard interactions
+
 ## 4.2.1
 
 * import lodash modules from main package